refactor: authRoutes
